+++ world/environment/__init__.py
@@ -1,1 +1,1 @@
-from . import pusher, objects
+from . import pusher_demo, pusher_haptic, pusher_haptic_depth, random_object_generator, rewards
