+++ world/environment/pusher_demo.py
@@ -1,0 +1,65 @@
+import numpy as np
+import pybullet as p
+
+from world.action.primitives import PushAction
+from world.environment.base import BaseEnv
+from world.physics.phys_net import HapticRegressor
+
+
+class PusherDemo(BaseEnv):
+    def __init__(self, config):
+        super().__init__(config)
+        self._observations_size = 21
+
+    @staticmethod
+    def get_model(config):
+        return HapticRegressor(batch_size=config["batch_size"],
+                               num_outputs=config["num_outputs"],
+                               action_kernel_size=config["action_kernel_size"],
+                               dropout=config["dropout"],
+                               lstm_units=config["lstm_units"],
+                               stateful_lstm=config["lstm_stateful"])
+
+    def get_observations(self, action: PushAction):
+        observations = list()
+
+        # set new position of the pusher w. r. t. the object
+        if self.scene["pusher"] is not None:
+            p.removeBody(self.scene["pusher"])
+
+        state_before = p.getBasePositionAndOrientation(self.object)
+        self.scene["pusher"], _, _ = self.setup_pusher(object_pos=state_before[0], action=action)
+        observations.append(state_before)
+
+        if action is not None:
+            # apply force on a pusher object
+            self.step_sim_with_force(action)
+            state_after = p.getBasePositionAndOrientation(self.object)
+            observations.append(state_after)
+
+            # wait more and get new observation
+            self.step_sim_with_force(action)
+            state_post_after = p.getBasePositionAndOrientation(self.object)
+            observations.append(state_post_after)
+
+        # flatten
+        observations = np.asarray([np.hstack(o) for o in observations])
+        observations = observations.reshape((1, -1))
+        return observations
+
+    def step(self, action: PushAction = None):
+        assert type(action) is PushAction and self.object is not None
+        observations, reward, done, info = list(), None, False, {}
+
+        # mock observations
+        info["haptic"] = self.rog.get_haptic_values()
+        observations = self.get_observations(action)
+        observations = np.asarray([np.hstack(o) for o in observations])
+        obs_flat = observations.reshape((1, -1))
+        act_flat = action.to_numpy().reshape((1, -1))
+        info["observations_numpy"] = (obs_flat, act_flat)
+
+        return observations, reward, done, info
+
+    def reset(self):
+        self.reset_sim()
